refactor: Transition 컴포넌트 커스텀훅으로 로직 분리
