--- frontend/src/lib/Router/location.ts
@@ -1,28 +1,0 @@
-interface LocationType {
-  path: string;
-  params: string[];
-  query: {
-    [key: string]: string;
-  };
-}
-
-export default (function () {
-  let location: LocationType = {
-    path: '',
-    params: [],
-    query: {},
-  };
-
-  const getLocation = () => {
-    return { ...location };
-  };
-
-  const setLocation = (newLocation: LocationType) => {
-    location = newLocation;
-  };
-
-  return {
-    getLocation,
-    setLocation,
-  };
-})();
